separated the generation of users for training the neural network into a separate directory, and also corrected errors related to changing the names of variables
