+++ src/ml/GenerationUsers/С_generation_fake_users.py
@@ -1,9 +1,7 @@
-from C_Assessment import Assessment
+from src.ml.GenerationUsers.C_Assessment import Assessment
 from typing import Dict, Any
 import pandas as pd
 import random
-import os
-
 
 
 class GitHubUserGenerator:
